% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bg-main.R
\name{bg_dvs}
\alias{bg_dvs}
\title{Inference for Bernoulli-Gaussian time varying parameter model with dynamic variable selection (! number of covariates should be p > 2). Several settings are available and both Variational Bayes and MCMC can be used.}
\usage{
bg_dvs(
  y,
  X,
  hyper,
  meanField = "full",
  options = list(sv = 1, smooth = 0),
  initialize = "null",
  Tol_Par = 0.01,
  Trace = 0
)
}
\arguments{
\item{y}{vector of observations (response variable)}

\item{X}{design matrix (\code{n x p})}

\item{hyper}{list of hyperparameters (see paper for details)}

\item{meanField}{type of mean-field factorization (see paper for details). Only "full" is available for now. "group", "joint" will be available soon. If group, group indicators should be given in the \code{hyper} list. If \code{meanField = 'null'}, MCMC is run (slow) and number of posterior samples \code{ndraws} and burn-in \code{nburn} should be provided in \code{hyper}.}

\item{options}{list containing two fields. \code{sv=0} (no stochastic volatility) \code{sv=1} (stochastic volatility, default) and \code{smooth=0} (no smoothing inclusion probabilities, default) \code{sv=1} (smoothing inclusion probabilities) (see the paper for details). Currently not all the options are supported by all the methods.}

\item{initialize}{default "null", that is \beta=0. If "OLS", OLS estimates are used to initialize \beta}

\item{Tol_Par}{tolerance in the relative variation of parameters to detect convergence in Variational Bayes}

\item{Trace}{if 1, print progress of the algorithm}
}
\value{
Variational parameters of the approximated posterior distributions according to the selected method (see the paper for details, the names of the variables match). If MCMC is used, posterior samples after burnin are returned.
}
\description{
Inference for Bernoulli-Gaussian time varying parameter model with dynamic variable selection (! number of covariates should be p > 2). Several settings are available and both Variational Bayes and MCMC can be used.
}
