\name{BGdvs-package}
\alias{BGdvs-package}
\alias{BGdvs}
\docType{package}
\title{
\packageTitle{BGdvs}
}
\description{
\packageDescription{BGdvs}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BGdvs}
\packageIndices{BGdvs}
~~ An overview of how to use the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{BGdvs}

Maintainer: \packageMaintainer{BGdvs}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
