% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{generateData}
\alias{generateData}
\title{Generate data from the model and the latent coefficients (time-varying parameters) with dynamic sparsity}
\usage{
generateData(
  n,
  p_in,
  p_dyn,
  p_low,
  p_null,
  lam,
  delta,
  thresh = 0.5,
  eta2 = 0.1,
  s2 = 0.25,
  nu2 = NULL
)
}
\arguments{
\item{n}{length of the series}

\item{p_in}{number of coefficients always included in the model}

\item{p_dyn}{number of coefficients dynamically included in the model}

\item{p_low}{number of coefficients included in the model only for a short period (low signal)}

\item{p_null}{number of coefficients always zero}

\item{lam}{average length of non-zero periods for \code{p_dyn} coefficients (should have same length!)}

\item{delta}{average length of non-zero period in the low signal case \code{p_low} coefficients (should have same length!)}

\item{thresh}{threshold for a parameter to be non-zero#'}

\item{eta2}{variance of the random walk process for \beta}

\item{s2}{variance of the error term}

\item{nu2}{if \code{NULL} (default) homoskedastic errors are considered. If a value is assigned, then stochastic volatility is included and \code{nu2} is the variance of the random walk process}
}
\value{
\code{y}: simulated observations

\code{X}: simulated design matrix \code{n x p} (independent standard normally distributed columns)

\code{beta}: the simulated latent coefficients \beta (a matrix \code{p x n})

\code{gamma}: the simulated latent indicators \gamma (a matrix \code{p x n})

\code{s2}: variance of the errors
}
\description{
Generate data from the model and the latent coefficients (time-varying parameters) with dynamic sparsity
}
